requirements.txt
kafka-python
python-dotenv
clickhouse-connect
argparse
